business creation page

+++ SmallBuisnessFinder/Objects/BusinessObject.swift
@@ -14,15 +14,15 @@ class BusinessObject {
     var name: String?
     var phoneNumber: String?
     var busDescription: String?
-    var latCoord: Int?
-    var longCoord: Int?
+    var latCoord: Double?
+    var longCoord: Double?
     var websiteLink: String?
     var following: Int?
     var followers: Int?
     var email: String?
     
     
-    init(name: String?, phoneNumber: String?, busDescription: String?, latCoord: Int?, longCoord: Int?, websiteLink: String?, following: Int?, followers: Int?) {
+    init(name: String?, phoneNumber: String?, busDescription: String?, latCoord: Double?, longCoord: Double?, websiteLink: String?, following: Int?, followers: Int?) {
         self.name = name
         self.phoneNumber = phoneNumber
         self.busDescription = busDescription
